Improve validations
- Move them to their own module.
- Make them generic over anything that can report errors.
